+++ src/github/responses.rs
@@ -1,17 +1,17 @@
 use serde::Deserialize;
 
-#[derive(Deserialize, PartialEq, Debug)]
+#[derive(Deserialize, PartialEq, Clone, Debug)]
 pub struct ActionsRuns {
     pub total_count: i32,
     pub workflow_runs: Vec<WorkflowRun>,
 }
 
-#[derive(Deserialize, PartialEq, Debug)]
+#[derive(Deserialize, PartialEq, Clone, Debug)]
 pub struct WorkflowRun {
     pub id: i32,
 }
 
-#[derive(Deserialize, PartialEq, Debug)]
+#[derive(Deserialize, PartialEq, Clone, Debug)]
 pub struct Repository {
     pub id: i32,
     pub name: String,
@@ -22,7 +22,7 @@ pub struct Repository {
     pub allow_merge_commit: bool,
 }
 
-#[derive(Deserialize, PartialEq, Debug)]
+#[derive(Deserialize, PartialEq, Clone, Debug)]
 pub struct StarredRepository {
     pub id: i32,
     pub name: String,
@@ -30,4 +30,23 @@ pub struct StarredRepository {
     pub html_url: String,
     pub description: Option<String>,
     pub language: Option<String>,
+}
+
+#[derive(Deserialize, PartialEq, Clone, Debug)]
+pub struct MyRepository {
+    pub id: i32,
+    pub name: String,
+    pub full_name: String,
+    pub html_url: String,
+    pub description: Option<String>,
+    pub language: Option<String>,
+    pub archived: bool,
+    pub visibility: RepositoryVisibility,
+}
+
+#[derive(Deserialize, PartialEq, Copy, Clone, Debug)]
+#[serde(rename_all = "snake_case")]
+pub enum RepositoryVisibility {
+    Public,
+    Private,
 }