+++ src/lib.rs
@@ -1,5 +1,1 @@
-#![deny(rust_2018_idioms)]
-
 pub mod github;
-
-pub use github::*;